pyproject.toml
[tool.poetry]
name = "manga-notif-bot"
version = "0.1.0"
description = ""
authors = ["Damego <danyabatueff@gmail.com>"]
readme = "README.md"
packages = [{include = "manga_notif_bot"}]

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = {extras = ["job-queue"], version = "^20.4"}
python-dotenv = "^1.0.0"
beautifulsoup4 = "^4.12.2"
beanie = "^1.20.0"
selenium = "^4.10.0"
webdriver-manager = "^3.9.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
